feat: criação das classes CheckInService, CheckInRepository e record DadosDetalharCheckIn e algumas modificações simples em outras classes

+++ src/main/java/Projeto/Gerenciador_Eventos/dto/DadosAtualizarEvento.java
@@ -8,7 +8,7 @@ import jakarta.validation.constraints.NotNull;
 public record DadosAtualizarEvento(	
 		
 		@NotNull
-		Integer idEvento,
+		Long idEvento,
 
 		String nomeEvento,
 		String descricaoEvento,
